pyproject.toml
[tool.poetry]
name = "generic-bug-strawberry"
version = "0.1.0"
description = ""
authors = ["Sarah Henkens <sarah.a.henkens@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
strawberry-graphql = "^0.57.1"
uvicorn = "^0.13.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
